fix(docker-image-deploy.yml): fix docker image tagging and pushing
The previous configuration was tagging the docker image with the "latest" tag, which is not ideal for versioning. Instead, we now tag the image with the commit SHA to ensure uniqueness. Additionally, we also tag the image as "latest" for convenience. This allows us to easily reference the latest version of the image when deploying.

The updated steps in the workflow now include:
- Building the docker image with the tag ${{ github.sha }}
- Tagging the image as $ECR_REGISTRY/$ECR_REPOSITORY:latest
- Pushing both the tagged image and the "latest" tagged image to the ECR registry.